fs: Move file reading helper from composefs-boot
Some bits in bootc will end up using this for the same
reason composefs-boot internals are.
